Realtor that Created House Can Delete or Update it

+++ src/home/home.controller.ts
@@ -1,4 +1,4 @@
-import { Body, Controller, Delete, Get, Param, ParseIntPipe, Post, Put, Query } from '@nestjs/common';
+import { Body, Controller, Delete, Get, Param, ParseIntPipe, Post, Put, Query, UnauthorizedException } from '@nestjs/common';
 import { PropertyType } from '@prisma/client';
 import { User, UserInfo } from 'src/user/decorators/user.decorator';
 import { CreateHomeDto, HomeResponseDto, UpdateHomeDto } from './dto/home.dto';
@@ -40,12 +40,20 @@ export class HomeController {
     }
 
     @Put(':id')
-    updateHome(@Param('id', ParseIntPipe) id: number, @Body() body: UpdateHomeDto) {
+    async updateHome(@Param('id', ParseIntPipe) id: number, @Body() body: UpdateHomeDto, @User() user: UserInfo) {
+        const realtor = await this.homeService.getRealtorByHomeId(id);
+
+        if (realtor.id !== user.id) throw new UnauthorizedException();
+
         return this.homeService.updateHome(id, body);
     }
 
     @Delete(':id')
-    deleteHome(@Param('id', ParseIntPipe) id: number) {
+    async deleteHome(@Param('id', ParseIntPipe) id: number, @User() user: UserInfo) {
+        const realtor = await this.homeService.getRealtorByHomeId(id);
+
+        if (realtor.id !== user.id) throw new UnauthorizedException();
+
         return this.homeService.deleteHomeById(id);
     }
